added tests

+++ bank-account-service/src/main/java/com/bank_account/event/TransferValidatedEvent.java
@@ -29,7 +29,7 @@ public class TransferValidatedEvent {
     private String initiatedBy;
 
     private boolean approved;
-    private double riskScore;
+    private BigDecimal riskScore;
     private List<FraudType> fraudTypes;
     private String rejectionReason;
